Try to address a Windowlicker timing problem
Windowlicker has a timing problem in version r268. It is discussed on
https://groups.google.com/forum/#!topic/growing-object-oriented-software/aEFI72rjn7w

Nat Pryce advises to use Google's WindowTester instead of Windowlicker.
Other users were successful in using a DEV version of Windowlicker. So I
am going this way.

Unfortunately the repository
  "https://ibuzz-csd.googlecode.com/hg/maven-repository"
is not available anymore as it was at timme committing badc8eb35ee423af8aa.
So i directly took the jars from
  https://github.com/edzhelyov/goos/tree/master/lib
and committed them to the lib directory.