Move ApplicationRunner, FakeAuctionServer, SingleMessageListener to integrationTest source set
This is to avoid a dependency loop between the integration test and
end-to-end test source set.

Updated the test-sets Gradle plugin to version 2.1.1, hoping that the
problem that sources of one test source set which use ones from a different
source set can now be compiled in IntelliJ IDEA. Unfortunately this still
is not the case. However, Gradle is able to compile and run it.